server/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3 # ASGI server for FastAPI
python-multipart==0.0.20 # for handling file uploads
sqlalchemy==2.0.41 # ORM for database interactions
psycopg2-binary==2.9.10 # allows communication with PostgreSQL
alembic==1.16.2 # for database migrations
python-jose # Creates and verifies JWT tokens (JSON Web Tokens) specifically "python-jose[cryptography]"
passlib==1.7.4 # for password hashing specifically "passlib[bcrypt]"
python-decouple==3.8
boto3==1.38.41 
pinecone-client==6.0.0
pinecone==7.2.0
PyPDF2==3.0.1
pytest==8.4.1
httpx==0.28.1
pytest-cov==6.1.1
email-validator==2.2.0
pytest-asyncio==1.0.0
pdfplumber==0.11.7
python-docx==1.2.0
langchain==0.3.26